pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Multi-agent pipeline to generate an MCP crash course with live research"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "replicate>=1.0.0",
    "markdown>=3.7",
    "weasyprint>=62.0",
    "pydub>=0.25.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
]

[project.scripts]
generate = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
